move coda out of torch.testing.__init__
